+++ index.css
@@ -61,9 +61,9 @@ table {
   width: 75%;
   align-items: center;
   justify-content: center;
-
 }
-tbody{
+
+tbody {
   width: 100%;
   display: flex;
   flex-direction: column;
@@ -125,15 +125,15 @@ input:focus {
   font-size: 0.9em;
 }
 
-.booksContainer tbody tr:nth-of-type(odd) {
-  background-color: var(--warmShade);
-}
-tr{
+tr {
   display: flex;
   gap: 30px;
   justify-content: space-evenly;
 }
 
+.booksContainer tbody tr:nth-of-type(odd) {
+  background-color: var(--warmShade);
+}
 
 .remove-btn:hover {
   box-shadow: 1px 1px 2px var(--warmShade);
