+++ modules/UI.js
@@ -1,35 +1,6 @@
-export class Store {
-  static getBooks = () => {
-    let books;
-    if (localStorage.getItem('books') === null) {
-      books = [];
-    } else {
-      books = JSON.parse(localStorage.getItem('books'));
-    }
-    return books;
-  }
-
-  static addBook = (book) => {
-    const books = Store.getBooks();
-    books.push(book);
-    localStorage.setItem('books', JSON.stringify(books));
-  }
-
-  static removeBook = (title) => {
-    const books = Store.getBooks();
-    books.forEach((book, index) => {
-      if (book.title === title) {
-        books.splice(index, 1);
-      }
-    });
-
-    localStorage.setItem('books', JSON.stringify(books));
-  }
-}
-
+import Store from './store.js';
 // UI class: Handles UI Tasks
-//import Store from "./store";
-export class UI {
+class UI {
   static displayBooks() {
     const books = Store.getBooks();
 
@@ -38,7 +9,6 @@ export class UI {
 
   static addBookToList(book) {
     const list = document.querySelector('#bookList');
-    
 
     const row = document.createElement('tr');
     row.classList.add('row');
@@ -74,3 +44,4 @@ export class UI {
     document.querySelector('#author').value = '';
   }
 }
+export default UI;
